autologin as separate component
