+++ pages/login.js
@@ -1,13 +1,17 @@
 import React from 'react';
 import Layout from 'components/layout/Layout';
 import LoginForm from 'components/forms/LoginForm';
+import AutoLogin from 'components/forms/AutoLogin';
 import { useRouter } from 'next/router';
 
 export default function LoginPage() {
   const { query } = useRouter();
+  if (query.hash) {
+    return <AutoLogin hash={query.hash} />;
+  }
   return (
     <Layout title="login" header={false} footer={false} center>
-      <LoginForm hash={query.hash} />
+      <LoginForm />
     </Layout>
   );
 }
